Move CLI to own module for fixer
